Fix PEP and add typing hints for most modules
Due to the refactoring from scarlet main, there are a number
of modules that just needed to be updated to the DM style guide
and have type hints added. The remaining files will need more
work, as they will also contain code changes to fix bugs and
make the API more uniform.