Refactor code and make old tests pass
The code was cleaned up and refactored,
including implementing an improved method of handling Parameters.
Type hints were added and tests were fixed to work with the
refactored code.

Note: there had been disagreements among the scarlet developers
as to how much test code should exist in scarlet, so the bulk
of the codebase is not yet tested, meaning there are likely more
bugs in the code. This commit is just a step towards having a
working package.