Pad all initial models
This is code that was meant to be turned on in the first place.
We always want to give sources room to grow, and it turns out
that for some faint sources with bright neighbors that were
initialized as a single pixel, due to the nature of the
monotonicity constraint, having a single pixel component
was causing some blends to fail.

So adding this already desired feature also fixed a bug
in initialization.