Pin astropy < 7
This is the same pin that the science pipelines uses because
astropy 7 changes its implementation of the display module.
We could just implement the old (and better) code in scarlet lite
except that Nate Lusts color code should be available soon and
that algorithm is even better, so we can try and holdout
for that.